sunxi: Use mmc_bootdev=3 for MMC2 boot
With DM_MMC the mmc devices are probed as per dt status
and eMMC can probed maximum device of 3, if all nodes like
mmc0, mmc1, mmc2 and mmc3 status are 'okay'.

So update mmc_bootdev to 3 and add boot order as 3, 2, 1, 0
devices. so-that it can to boot any identified device even
if the respective device status disabled or unsupported.

Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>